+++ src/core/process_management/gt_stringmonitoringproperty.cpp
@@ -27,3 +27,9 @@ GtStringMonitoringProperty::GtStringMonitoringProperty(const QString& ident,
 {
 
 }
+
+gt::PropertyFactoryFunction
+gt::makeStringMonitoringProperty(QString value)
+{
+    return makePropertyFactory<GtStringMonitoringProperty>(std::move(value));
+}